feat: modularized code into partials, fixed 'Find' and 'Replace' dialog layout.
