--- Program.cs
@@ -1,68 +1,0 @@
-/*
- * ognp (OG Notepad)
- * A faithful Notepad clone written in C#.
- *
- * © 2025 Jim Sines (raystanza).
- *
- * Licensed under GPL-3.0-or-later.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
-
-using System;
-using System.IO;
-using System.Windows.Forms;
-
-namespace ognp
-{
-    internal static class Program
-    {
-        [STAThread]
-        static void Main(string[] args)
-        {
-            ApplicationConfiguration.Initialize();
-
-            string? path = null;
-            int? line = null;
-
-            // Accept either: ognp.exe +123 file.txt  OR  ognp.exe file.txt +123
-            foreach (var a in args)
-            {
-                if (a.StartsWith("+", StringComparison.Ordinal) &&
-                    int.TryParse(a.AsSpan(1), out var n) && n > 0)
-                {
-                    line = n;
-                }
-                else
-                {
-                    path = a;
-                }
-            }
-
-            if (!string.IsNullOrWhiteSpace(path))
-            {
-                try
-                {
-                    path = Path.GetFullPath(path!);
-                }
-                catch (ArgumentException) { /* ignore bad paths */ }
-                catch (NotSupportedException) { /* ignore bad paths */ }
-                catch (PathTooLongException) { /* ignore bad paths */ }
-                catch (System.Security.SecurityException) { /* ignore bad paths */ }
-            }
-
-            Application.Run(new MainForm(path, line));
-        }
-    }
-}
